1. adding class .all on signup button, login form div, signup form div, addTodo form div, and todos div 2. adding a signup form jquery for a show and hide system 3. adding button for register and login instead 4. Tidying up the UserController file, TodoController file, and userRoutes file 5. Creating env-templates file with the environment variables
