+++ client/jquery.js
@@ -2,175 +2,70 @@
 
 let $todoContainer = $("#todos-container")
 let $todoTbody = $("#todos-tbody")
+
+let $login = $('#login')
 let $loginForm = $('#form-login')
+let $btnSignout = $('#signout-button')
+
+let $signup = $('#signup')
 let $signupForm = $('#form-signup')
-let $btnSignout = $('#button-signout')
+let $signupButton = $('#button-signup')
+let $loginInsteadButton = $('#button-loginInstead')
+
+let $addTodo = $('#form-addTodo-container')
 let $addTodoForm = $('#form-addTodo')
-let $getUrl = $('#get-url')
+
+let $googleSignin = $('#google-signin')
+let $addTodoButton = $('#add-todo-button')
+let $allPage = $('.all')
 
 if (localStorage.getItem("token") !== null) {
     $(document).ready(() => {
-        $loginForm.hide()
-        $signupForm.hide()
+        // $login.hide()
+        // $signup.hide()
+        $allPage.hide()
         $btnSignout.show()
+        $todoContainer.show()
         showTodo()
     })
+} else if (localStorage.getItem("signup") === 'on sign up') {
+    $allPage.hide()
+    $signup.show()
 } else {
     $(document).ready(() => {
-        $signupForm.hide()
-        $btnSignout.hide()
-        $todoContainer.hide()
-        $addTodoForm.hide()
+        $allPage.hide()
+        $login.show()
     })
 }
 
-const login = (email, password) => {
-    $.ajax({
-        method: 'post',
-        url: 'http://localhost:3000/login',
-        data: {"email": email, "password": password},
-        success: result => {
-            let token = result
-            localStorage.setItem("token", token)
-            $loginForm.hide()
-            $btnSignout.show()
-            $addTodoForm.show()
-            $todoContainer.show()
-            showTodo()
-        },
-        error: (jqxhr, status, error) => {
-            console.log(status, error)
-        }
-    })
-}
+$signupButton.on('click', () => {
+    localStorage.setItem("signup", "on sign up")
+    $allPage.hide()
+    $signup.show()
+})
 
-const showTodo = () => {
-    $.ajax({
-        method: 'get',
-        url: 'http://localhost:3000/todos',
-        headers: {
-            "token": localStorage.getItem("token")
-        },
-        success: (data) => {
-            $('#form-login').hide()
-            if (data.length) {
-                $todoTbody.empty()
-                data.forEach(todo => {
-                    let $todoItem =  $(`<tr><td>${todo.title}</td><td>${todo.description}</td><td>${todo.due_date}</td></tr>`)
-                    let $btnDelete = $(`<td><button class="btn btn-primary" >Delete</button></td>`)
-                    let $btnCheck = $(`<td><input type="checkbox" class="form-check-input ml-2" ${todo.status === 'Completed' ?  'checked' : ''}></td>`)
-                    $btnDelete.on('click', () => {
-                        deleteTodo(todo.id)
-                    })
-                    $btnCheck.on('click', (e) => {
-                        checkTodo(todo, e.target.checked)
-                    })
-                    $todoItem.append($btnDelete)
-                    $todoItem.prepend($btnCheck)
-                    $todoTbody.append($todoItem)
-                })
-            } else {
-                $todoTbody.empty()
-            }
-        },
-        error: (jqxhr, status, error) => {
-            console.log(error)
-        }
-    })
-}
+$loginInsteadButton.on('click', () => {
+    localStorage.removeItem("signup")
+    $allPage.hide()
+    $login.show()
+})
 
-const addTodo = todo => {
-    $.ajax({
-        method: 'post',
-        url: `http://localhost:3000/todos`,
-        headers: {
-            "token": localStorage.getItem("token")
-        },
-        data: todo,
-        // data: {
-        //     "title": todo.title,
-        //     "description": todo.description,
-        //     "due_date": todo.due_date
-        // },
-        success: () => {
-            console.log('todo created')
-            $addTodoForm[0].reset()
-            showTodo()
-        }
-    })
-}
-
-const deleteTodo = id => {
-    $.ajax({
-        method: 'delete',
-        url: `http://localhost:3000/todos/${id}`,
-        headers: {
-            "token": localStorage.getItem("token")
-        },
-        success: () => {
-            showTodo()
-        }
-    })
-}
-
-const checkTodo = (todo, status) => {
-    $.ajax({
-        method: 'put',
-        url: `http://localhost:3000/todos/${todo.id}`,
-        headers: {
-            "token": localStorage.getItem("token")
-        },
-        data: {
-            "title" : todo.title,
-            "description" : todo.description,
-            "status" : status ? "Completed" : "Not Started",
-            "due_date" : todo.due_date
-        },
-        success: () => {
-            showTodo()
-        }
-    })
-}
-let $urlButton
-function onSignIn(googleUser) {
-    let id_token = googleUser.getAuthResponse().id_token;
-    $.ajax({
-        method: 'post',
-        url: 'http://localhost:3000/googleLogin',
-        data: { "token": id_token },
-        success: data => {
-            console.log('CIEEEE SIGN IN')
-            $urlButton = $(`<a href=${data.url} class="btn btn-primary">Integrate with google calendar</a>`)
-            localStorage.setItem("token", data.newToken)
-            $getUrl.empty()
-            $getUrl.append($urlButton)
-            $loginForm.hide()
-            $todoContainer.show()
-            $btnSignout.show()
-            $addTodoForm.show()
-            showTodo()
-        }
-    })
-}
-
-function signOut() {
-    let auth2 = gapi.auth2.getAuthInstance();
-    auth2.signOut().then(function () {
-      localStorage.removeItem("token")
-      $signupForm.hide()
-      $todoContainer.hide()
-      $loginForm.show()
-      $btnSignout.hide()
-      $addTodoForm.hide()
-      console.log('User signed out.');
-    });
-}
+$signupForm.on('submit', (e) => {
+    e.preventDefault()
+    let $firstName = $('#firstname-signup').val()
+    let $lastName = $('#lastname-signup').val()
+    let $email = $('#email-signup').val()
+    let $password = $('#password-signup').val()
+    register($firstName, $lastName, $email, $password)
+    $signupForm[0].reset()
+})
 
 $loginForm.on('submit', (e) => {
     e.preventDefault()
     let $email = $('#email-login').val()
     let $password = $('#password-login').val()
     login($email, $password)
+    $loginForm[0].reset()
 })
 
 $addTodoForm.on('submit', (e) => {
@@ -184,7 +79,27 @@ $addTodoForm.on('submit', (e) => {
         "due_date": $todoDueDate
     }
     addTodo(newTodo)
+    $allPage.hide()
+    $todoContainer.show()
+    $btnSignout.show()
 })
+
+
+$googleSignin.on('click', () => {
+    $allPage.hide()
+    $btnSignout.show()
+    $todoContainer.show()
+    onSignIn()
+})
+
+$addTodoButton.on('click', () => {
+    $allPage.hide()
+    $addTodo.show()
+})
+
+
+
+
