+++ server/controllers/TodoController.js
@@ -1,8 +1,8 @@
 const { Todo } = require('../models')
-const { google } = require('googleapis')
-const { client, acctoken } = require('./UserController')
+// const { google } = require('googleapis')
+// const { client, acctoken } = require('./UserController')
 const credentials = require('../credentials')
-const { authorize, insertEvents } = require('../index')
+const { authorize, insertEvents } = require('../googleCal')
 
 class TodoController {
     static findAll = (req, res, next) => {
@@ -37,22 +37,8 @@ class TodoController {
                         "dateTime" : new Date(todo.due_date)
                     }
                 }
-                // res.status(201).json({ msg: "New Todo has been created", newTodo })
                 let result = authorize(credentials,event, insertEvents)
-                console.log(result)
                 res.status(201).json({ msg: "New Todo has been created", newTodo })
-
-                // calendar.events.insert({
-                //     calendarId: 'primary',
-                //     resource: event
-                // }, (err, result) => {
-                //     if (result) {
-                //         console.log(result)
-                //         res.status(201).json({ msg: "New Todo has been created", newTodo })
-                //     } else {
-                //         console.log(err)
-                //     }
-                // })
             })
             .catch(err => {
                 if (err.errors) {
