+++ server/env-templates
@@ -1,0 +1,6 @@
+PORT=
+DB_USERNAME=
+DB_PASSWORD=
+JWT_SECRET=
+CLIENT_ID=
+CLIENT_SECRET=
